Update mw-node-qunit tests
Refresh the tests to express how v6 works.
A new function setUp is added to handle the case where tests do
not want to explicitly express dependencies. It is invoked by
the --globals keyword and will be used by Popups

Support for the `this.sandbox` mechanism is removed - use of
sinon.sandbox.create() is preferred going forward.